client/requirements.txt
fusepy
secrets
pycryptodome

